new  chamges

+++ experiment_12 (1)/experiment_12/src/main/java/com/example/experiment_12/EmployeeRepository.java
@@ -1,8 +1,7 @@
 package com.example.experiment_12;
 
+import com.example.experiment_12.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository
-        extends JpaRepository<Employee, Integer> {
-
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 }
